Make include_package_data=True for `pyproject.toml` configs
There is frequent an opinion in the community that
`include_package_data=True` is a better default
(and a quality of life improvement).

Since we are migrating to a new configuration file, this change can
be implemented in a backward compatible way
(to avoid breaking existing packages):

- Config from `setup.cfg` defaults to `include_package_data=False`
- Config from `pyproject.toml` defaults to `include_package_data=True`

This also takes advantage that `ini2toml` (the provided library for
automatic conversion between `setup.cfg` and `pyproject.toml`) will
backfill `include_package_data=False` when the field is missing.